shortlink: 基本实现短链接跳转功能

+++ programe/src/main/java/com/project/service/impl/ShortLinkServiceImpl.java
@@ -9,6 +9,8 @@ import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.project.common.convention.exception.ClientException;
 import com.project.dao.entity.ShortLinkDO;
+import com.project.dao.entity.ShortLinkGoToDO;
+import com.project.dao.mapper.ShortLinkGoToMapper;
 import com.project.dao.mapper.ShortLinkMapper;
 import com.project.dto.req.ShortLinkCreateReqDTO;
 import com.project.dto.req.ShortLinkPageReqDTO;
@@ -18,10 +20,15 @@ import com.project.dto.resp.ShortLinkCreateRespDTO;
 import com.project.dto.resp.ShortLinkPageRespDTO;
 import com.project.service.ShortLinkService;
 import com.project.util.HashUtil;
+import jakarta.servlet.ServletRequest;
+import jakarta.servlet.ServletResponse;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.AllArgsConstructor;
 import org.redisson.api.RBloomFilter;
 import org.springframework.stereotype.Service;
 
+import java.io.IOException;
 import java.util.List;
 import java.util.Map;
 
@@ -33,6 +40,7 @@ import java.util.Map;
 @AllArgsConstructor
 public class ShortLinkServiceImpl extends ServiceImpl<ShortLinkMapper, ShortLinkDO> implements ShortLinkService {
     private final RBloomFilter<String> shortLinkCachePenetrationBloomFilter;
+    private final ShortLinkGoToMapper ShortLinkGoToMapper;
 
 
     @Override
@@ -55,6 +63,7 @@ public class ShortLinkServiceImpl extends ServiceImpl<ShortLinkMapper, ShortLink
 
         baseMapper.insert(shortLinkDO);
         shortLinkCachePenetrationBloomFilter.add(shortUrl);
+        ShortLinkGoToMapper.insert(ShortLinkGoToDO.builder().shortUrl(shortUrl).gid(requestParam.getGid()).build());
 
 
         return ShortLinkCreateRespDTO
@@ -102,14 +111,64 @@ public class ShortLinkServiceImpl extends ServiceImpl<ShortLinkMapper, ShortLink
         if (shortLinkDO == null) {
             throw new ClientException("短链接不存在");
         }
-        //由于可能要修改分组，而分组gid是t_link的分片键，所以只能先删除再插入
-        baseMapper.deleteById(shortLinkDO.getId());
-        shortLinkDO.setGid(requestParam.getGid());
+
         shortLinkDO.setDescribe(requestParam.getDescribe());
         shortLinkDO.setValidDateType(requestParam.getValidDateType());
         shortLinkDO.setValidDate(requestParam.getValidDate());
         shortLinkDO.setOriginUrl(requestParam.getOriginUrl());
-        baseMapper.insert(shortLinkDO);
+        //由于可能要修改分组，而分组gid是t_link的分片键，所以只能先删除再插入
+        if (!requestParam.getGid().equals(requestParam.getOriginGid())) {
+            //说明分组发生了变化
+            baseMapper.deleteById(shortLinkDO.getId());
+            shortLinkDO.setGid(requestParam.getGid());
+            baseMapper.insert(shortLinkDO);
+            //更新短链接的goto表
+            ShortLinkGoToDO shortLinkGoToDO = ShortLinkGoToDO.builder()
+                    .gid(requestParam.getGid()).build();
+            LambdaQueryWrapper<ShortLinkGoToDO> eq1 = Wrappers.lambdaQuery(ShortLinkGoToDO.class)
+                    .eq(ShortLinkGoToDO::getShortUrl, shortLinkDO.getFullShortUrl());
+            ShortLinkGoToMapper.update(shortLinkGoToDO, eq1);
+        }else {
+            //这里由于gid是分片键，更新的实体不能携带gid，所以要先将实体gid设置为null，然后再更新
+            shortLinkDO.setGid(null);
+            baseMapper.updateById(shortLinkDO);
+        }
+
+    }
+
+    @Override
+    public void shortLinkRedirect(String shortUri, ServletRequest request, ServletResponse response)  {
+        //由于短链接的分片键是gid，查询尽量用gid查询，不然就会查询所有的表，
+        //所以这里我创建了一个新的表，用来存储短链接和gid的对应关系goto
+        String protocol= ((HttpServletRequest) request).getScheme();
+        String shortUrl = protocol +request.getServerName() + "/" + shortUri;
+        LambdaQueryWrapper<ShortLinkGoToDO> eq = Wrappers.lambdaQuery(ShortLinkGoToDO.class)
+                .eq(ShortLinkGoToDO::getShortUrl, shortUrl);
+        ShortLinkGoToDO shortLinkGoToDO = ShortLinkGoToMapper.selectOne(eq);
+        if (shortLinkGoToDO == null) {
+            //如果短链接不存在，返回404
+            ((HttpServletResponse)response).setStatus(404);
+            return;
+        }
+        //如果短链接存在，重定向到原始链接
+        LambdaQueryWrapper<ShortLinkDO> eq1 = Wrappers.lambdaQuery(ShortLinkDO.class)
+                .eq(ShortLinkDO::getGid, shortLinkGoToDO.getGid())
+                .eq(ShortLinkDO::getFullShortUrl, shortUrl)
+                .eq(ShortLinkDO::getEnableStatus, 0)
+                .eq(ShortLinkDO::getDelFlag, 0);
+        ShortLinkDO shortLinkDO = baseMapper.selectOne(eq1);
+        if (shortLinkDO == null) {
+            //如果短链接不存在，返回404
+            ((HttpServletResponse) response).setStatus(404);
+            return;
+        }
+
+        try {
+            ((HttpServletResponse)response).sendRedirect(shortLinkDO.getOriginUrl());
+        } catch (IOException e) {
+            throw new RuntimeException(e);
+        }
+
     }
 
 }
